refactor: move excel upload handler to api package
